requirements.txt
python-telegram-bot>=22.0
google-genai>=1.12.0
python-dotenv>=1.1.0 